+++ .gitignore
@@ -1,12 +1,14 @@
-.vscode/
-venv*/
-testdir/
-testdir_persistent/
-__pycache__/
-mcserverwrapper/test/temp/
-*.egg-info/
-.pytest_cache/
-examples/temp/
-dist/
+*\__pycache__
+*\.pytest_cache
+*\.mypy_cache
+*\.ruff_cache
+*\.vscode
+*\venv*
+*\dist
+*\testdir
+*\testdir_persistent
 *.pyc
+*.egg-info
+src/test/temp
+examples/temp
 password.txt
