--- mcserverwrapper/__init__.py
@@ -1,7 +1,0 @@
-"""Export Wrapper class"""
-
-from mcserverwrapper.src import Wrapper
-
-__exports__ = [
-    Wrapper
-]
