--- mcserverwrapper/src/server/__init__.py
@@ -1,13 +1,0 @@
-"""Export server classes"""
-
-from .server_builder import ServerBuilder
-from .base_server import BaseServer
-from .paper_server import PaperServer
-from .vanilla_server import VanillaServer
-
-__exports__ = [
-    ServerBuilder,
-    BaseServer,
-    PaperServer,
-    VanillaServer
-]
