+++ src/mcserverwrapper/error.py
