+++ src/test/integration_tests/test_vanilla.py
@@ -1,17 +1,18 @@
 """Module containing tests for Vanilla servers"""
 
 import os
+from datetime import datetime, timedelta
 from random import randint
 from time import sleep
-from datetime import datetime, timedelta
 
 import pytest
 
-from mcserverwrapper import Wrapper
-from mcserverwrapper.src import error
-from ..helpers.common_helper import assert_port_is_free, download_file, connect_mineflayer, setup_workspace
-from ..helpers.vanilla_helper import run_vanilla_test, run_vanilla_test_url
-from ..testable_thread import TestableThread
+from mcserverwrapper import Wrapper, error
+from test.helpers.common_helper import (assert_port_is_free,
+                                        connect_mineflayer, download_file,
+                                        setup_workspace)
+from test.helpers.vanilla_helper import run_vanilla_test, run_vanilla_test_url
+from test.testable_thread import TestableThread
 
 def test_all(jar_version_tuple):
     """Tests all of the vanilla minecraft versions"""
