+++ src/test/module_tests/test_server_properties_helper.py
@@ -1,10 +1,10 @@
 """Test the server_properties_helper methods"""
 
-import pathlib
 import os
+import pathlib
 
-from mcserverwrapper.src.mcversion import McVersion, McVersionType
-from ...src import server_properties_helper as sph
+from mcserverwrapper import server_properties_helper as sph
+from mcserverwrapper.mcversion import McVersion, McVersionType
 
 def test_get_mixed_params():
     """Tests the helper with mixed params"""
